Make heap generic over *d*
The `Dary` trait is added as a marker everywhere where necessary, and
`BinaryHeap` is renamed to `DaryHeap`. The *logic* is not yet updated,
so regardless of `Dary` the heap will still be binary for now.

The feature `specialization_extend` was removed because it is currently
not possible by relying on the `min_specialization` compiler feature; it
is possible using `specialization` but this is a compiler feature that
is known to potentially generate unsound code.