Add d-ary logic
Children/parents are now calculated based on *d*; tests are added to
assert heap validity for *d* in 2, 3, 4, 8. These are also added as `Dx`
enums and type aliases in general. The heuristics to decide between
rebuilding and extending in `append` have *not* been re-evaulated yet
for *d* unequal to 2; this is a potential performance issue but does not
affect correctness.