Partially optimize `append`
This method chooses between two strategies for merging heaps, being
1. Pushing all elements of the smaller heap onto the larger one,
2. Merging both inner `Vec`s and then rebuilding the heap.
Strategy 1 is generally cheaper for appending a smaller heap, while 2 is
more appropriate when the heaps are similar in size. Calculating the
cut-off point was done without taking *d* into account, this commit
changes that to calculating the cut-off point based on the upper bound
of the number of comparisons that have to be performed.

This commit also adds a benchmark to aid in determining whether the
cut-off point between the two strategies is in the correct place.
Taking *d* into account as it is done now improves performance for 4-
and 8-heaps, but degrades 3-heap performance due to the use of integer
2-based logarithms; floating-point logarithms are not available in
`no_std` without implementing it ourselves or depending on an external
crate. However, in all heaps benchmarked, the cut-off point for
switching from strategy 1 to 2 happens too early: strategy 2 is applied
to heap lengths that could be merged faster with strategy 1 (under
benchmark conditions). This should be addressed in the future, perhaps
by deciding based on the average number of comparisons instead of
the worst case number.