Safeguards against nullary heap
Unary and nullary heaps are presently constructable, but have
respectively little and no use. Unary heaps, while technically valid,
are essentially sorted vectors with poor performance. Nullary heaps
can be constructed, but panic upon adding more than one element because
various code divides by the arity and dividing by zero panics. Several
code paths in unsafe code rely on the arity to be greater than zero, but
these can fortunately not be reached as no nullary heaps with more
than one item can be constructed safely by code outside the crate
without causing a divide-by-zero panic.

To ensure the unsound code paths do not become reachable in the future,
assertions (that *d* != 0) were added at the top of all functions that
use the arity as a number in their calculations. The `arity` macro was
also altered to make it more difficult to construct a nullary heap, but
it remains possible by implementing the `Arity` trait manually.

An alternative solution is to prevent the construction of nullary heaps.
This can be done at compile-time by using `NonZeroUsize` instead of
simply `usize` in the definition of `Arity`, or at runtime by placing
the *d* != 0 assertions at every location where a nullary heap can be
constructed. The former solution is maybe elegant but less ergonomic
as (1) it requires `.get()` everywhere where it's used and (2) on
current stable it can only be unsafely constructed in a const context.
The latter is possible, but would result in those assertions in addition
to the currently added assertions, which makes for a rather large amount
of assertions overall.