Establish stability guidelines and set MSRV
The minimum supported Rust version is set to 1.41.0, because this
version is required to implement `From<DaryHeap<T, D>>` for `Vec<T>`.
This version is added to CI to make sure the crate actually compiles.
However, tests are disabled (just the build remains) because the
dev-dependencies may have higher MSRVs.

The facts that the features unstable(_nightly) are not stable and that
nullary heaps are not considered stable are also documented.