Change serde (de)serialization format
When using a simple derive annotation, the heap property is not ensured
when constructing via `deserialize`, and the serde representation is not
the same as that implemented for `std::collections::BinaryHeap`. By
default, the struct and field names are also included in the serde
format, while a standard library type like `BinaryHeap` is simply a
sequence without anything named.

Solving either the heap property or the data representation issue is
possible by adding a serde attribute to the struct and keeping the
derive macros, but solving both seems impossible with just the derive
macros. Therefore, a custom implementation was added. For serialization
the inner `Vec` is simply serialized; for deserialization first the
inner `Vec` is deserialized, and then `heap.rebuild()` (directly or
indirectly via `From<Vec<T>> for DaryHeap<T, D>`) is used to assert the
heap property. If the serialized data already conformed to the heap
property, this is cheap (just n / d comparisons, no swaps necessary).
The rebuilding of the heap can yield a different inner representation
than that of `std::collections::BinaryHeap`, as there `BinaryHeap::push`
is used repeatedly to construct the heap.

The function `deserialize_in_place` is also implemented, where the heap
is first cleared before adding new elements to make sure that on error
the heap does not contain elements from the previous heap. This is again
to ensure identical behavior to the deserialization as implemented for
`std::collections::BinaryHeap`. The same caveat of `push` versus
`rebuild` applies, though.