Document maximum heap capacity
Due to code like D::D * pos + 1 in several places, overflows can happen
when the length of a heap exceeds (usize::MAX - 1) / D::D. This could
relatively easily be solved by using saturating multiplication and
addition, but this has a noticeable performance impact. Therefore, for
now, simply a warning is added to the documentation.