Synchronize with Rust 1.63.0
The methods `try_reserve` and `try_reserve_exact` are stabilized, but
require a MSRV of 1.57.0 (due to the underlying functions of `Vec` of
the same name), and are therefore only available when the `extra`
feature is enabled.

Other changes are to documentation comments and the random number
generator used for benchmarks. The location of tests was changed
upstream; the synchronization script was updated to account for this.