Synchronize with Rust 1.68.0
Updates in tests, and `PeekMut` prefers leaking heap elements over
leaving them in an inconsisitent state when it itself is leaked.

The crash test dummy from the standard library is copied as well, but
its special implementation of a RNG is not, as thread_rng() works fine
for our purposes at the moment.

The synchronization script is updated for a different path for the main
source code and the additional crash test dummy.