Synchronize with Rust 1.70.0
The `retain` method is stabilized. It doesn't require raising the MSRV,
so it's now available without any feature flags.

`RebuildOnDrop` is refactored out so it can also be used in the new
`extend` implementation. The old `extend` implementation in the standard
library used specialization to improve performance in some cases, which
we couldn't use, but the new one does not, so performance of this crate
should now be on par again with the standard library.

`Default` is now implemented for `IntoIter`. In the standard library,
this uses the `Default` implementation of `vec::IntoIter`, but since
that was only added in Rust 1.70.0 as well, we construct an identical
empty `vec::IntoIter` manually.

Finally, some unnecessary imports were removed. Less than in the
standard library, because we are still on edition 2018.