Revert "Update criterion dependency"
This reverts commit 60349273c8be74d643ec83ead1daa6e6a13cdaf7.

A transient dependency of the updated criterion is `clap = "^4"`, which
specifies a `rust-version` larger than our MSRV. This results in a Cargo
error, even when compiling the crate without using the dev-dependencies.
In order to keep the current MSRV, revert the update for now and figure
out if there's a way to do this properly later.

Even though it is only
a dev-dependency, this results in a Cargo error when not using
dev-dependencies. For now, let's revert it