Update criterion dependency
There doesn't seem to be an easy way to avoid raising the MSRV when
updating the dev-dependency to 0.5, so update it to 0.4 for now. A
reduction of compilation time should already be possible this way.
Moving the benchmarks to a separate crate like for fuzzing could be
considered for the future.