Allow non-beneficial lints
Rationales:
For `unexpected_cfgs`: we use the nonstandard `fuzzing` and `docsrs`
features, and would like to keep using them without adding a build
script.
For `non_local_definitions`: all `serde`-related definitions are
contained within a module for ease of feature-flag switching, and since
this is the only non-test module and not in a different file, the lint
is otherwise unnecessary.